update name of the test file to match the classes and add a fe extra tests
